requirements.txt
streamlit
pandas
pyarrow
fastparquet
langchain
langchain-openai
langchain-core
langchain-community
langchain-text-splitters
ragas
ipywidgets
pillow
python-dotenv
pypdf
tiktoken
ragas
plotly